+++ source/rafcon/core/state_elements/data_port.py
@@ -248,8 +248,6 @@ class DataPort(StateElement):
         if default_value is not None:
             # If the default value is passed as string, we have to convert it to the data type
             if isinstance(default_value, str):
-                if len(default_value) > 1 and default_value[0] == '$':
-                    return default_value
                 if default_value == "None":
                     return None