+++ tests/assets/scripts/hello_world_write_global_variable.py
@@ -4,7 +4,7 @@ def execute(self, inputs, outputs, gvm):
     self.logger.info("Executing performance test hello world sleep state ...")
     # import time
     # time.sleep(0.2)
-    print("+++++++++++++++++++++++", id(threading.currentThread()), "+++++++++++++++++++++++")
+    print("+++++++++++++++++++++++", id(threading.current_thread()), "+++++++++++++++++++++++")
     print("+++++++++++++++++++++++", threading.current_thread(), "+++++++++++++++++++++++")
     print("+++++++++++++++++++++++", threading.current_thread().name, "+++++++++++++++++++++++")
     if not gvm.variable_exist("test_counter"):