+++ tests/assets/unit_test_state_machines/start_from_here_test/develop_test_libraries/example_set_some_data_2/ExecutionState_3_NAYMSQ/script.py
@@ -1,0 +1,4 @@
+
+def execute(self, inputs, outputs, gvm):
+    outputs["output_2"] = 44
+    return "success"