+++ tests/assets/unit_test_state_machines/start_from_here_test/develop_test_libraries/execution_state/script.py
@@ -1,0 +1,6 @@
+
+def execute(self, inputs, outputs, gvm):
+    outputs["output_2"] = "some manual output"
+    outputs["output_3"] = [0, 1, 2, 3, 4]
+    outputs["output_4"] = 7
+    return "success"