+++ tests/gui/test_wt_environment.py
@@ -7,6 +7,7 @@ import threading
     'RIGHT_BAR_HIDDEN': False,
     'CONSOLE_HIDDEN': False,
 }}], indirect=True, ids=["with visible sidebars"])
+@pytest.mark.filterwarnings("ignore::DeprecationWarning")
 def test_thread_observer_creation_list(gui):
     """ Test that the worker thread (i.e. the thread calling this function) is not the gui thread
     """
@@ -14,10 +15,10 @@ def test_thread_observer_creation_list(gui):
     from rafcon.core.states.hierarchy_state import HierarchyState
     from rafcon.core.state_machine import StateMachine
 
-    print("WT_ident: ", threading.currentThread().ident)
+    print("WT_ident: ", threading.current_thread().ident)
     print("CORE_singleton_init_thread_ident: ", gui.core_singletons.thread_identifier)
     print("GUI_singleton_init_thread_ident: ", gui.singletons.thread_identifier)
-    assert gui.singletons.thread_identifier != threading.currentThread().ident
+    assert gui.singletons.thread_identifier != threading.current_thread().ident
     main_window_controller = gui.singletons.main_window_controller
     menubar_ctrl = main_window_controller.menu_bar_controller
     gui(menubar_ctrl.on_new_activate, None)